Update Install
Removal of sudo user: Sudo was not installed anyway, and root is good enough.
Keyboard layout and time zone made as variables to make them easy to change.
EFI set to 128MiB: Probably good enough for Hyper-V, not sure.
base-devel removed: Base-devel package not necessary.
Replaced echo in favor of printf: printf is native to Bash, so faster. Not like it needs to be faster.
Removed multilib: Useless for testing in VMs.
Sed for locale.gen: Seeking desired line instead of a fixed line number is easier and requires no maintenance.
Variables are now enclosed in curly brackets to ensure the right variables are being used.